+++ activesupport/lib/active_support/inflector/methods.rb
@@ -172,9 +172,7 @@ module ActiveSupport
     #   titleize('raiders_of_the_lost_ark')                      # => "Raiders Of The Lost Ark"
     #   titleize('string_ending_with_id', keep_id_suffix: true)  # => "String Ending With Id"
     def titleize(word, keep_id_suffix: false)
-      humanize(underscore(word), keep_id_suffix: keep_id_suffix).gsub(/\b(?<!\w['’`])[a-z]/) do |match|
-        match.capitalize
-      end
+      I18n.translate("string.format.titleize", word: word, keep_id_suffix: keep_id_suffix)
     end
 
     # Creates the name of a table like Rails does for models to table names.